# docs/swagger.yaml
basePath: /
definitions:
...
  internal_http-server_handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_http-server_handlers.permission:
    properties:
      granted_by:
        type: string
      permission:
        type: string
      position_id:
        type: string
    type: object
  internal_http-server_handlers.permissions:
    properties:
      permissions:
        items:
          $ref: '#/definitions/internal_http-server_handlers.permission'
        type: array
      position_id:
        type: string
    type: object
...
  internal_http-server_handlers.positionCreate:
    properties:
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/internal_http-server_handlers.permission'
        type: array
    type: object
...
  internal_http-server_handlers.updatePosition:
    properties:
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/internal_http-server_handlers.permission'
        type: array
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: API Server for challenge service
  title: Acheer API
  version: "1.0"
paths:
...
  /api/user/perm/:id:
    put:
      consumes:
      - application/json
      description: Remove Permissions From Position
      operationId: remove-permissions
      parameters:
      - description: position id with permissions
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_http-server_handlers.permissions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove Permissions From Position
      tags:
      - positions
  /api/user/position/:
    post:
      consumes:
      - application/json
      description: create models.Position
      operationId: create-position
      parameters:
      - description: position info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_http-server_handlers.positionCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Position type
      tags:
      - positions
  /api/user/position/:id:
    put:
      consumes:
      - application/json
      description: Update position by id
      operationId: update-position
      parameters:
      - description: position type
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_http-server_handlers.updatePosition'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update position by id
      tags:
      - positions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"